Stop using shorts and chars where ints are better
C implicitly converts variables to int all the time, so using chars
and shorts everywhere will just result in a ton of unnecessary
casting.

This conversion only extends to local variables and function
parameters: state variables still use the smallest available types,
as state is optimized for size rather than performance.

This eliminates all warnings generated when compiling clownmdemu with
MSVC 12.0/Visual Studio 6.

Sidenote: it's so cool to finally be able to have MSVC 12.0 compile
my code as C for once. Usually, I program in a common subset of C99
and C++98, and rely on compiling my code as C++ in order for building
with MSVC 12.0 to work. With C89, it can just compile as C instead.

+++ vdp.h
@@ -111,11 +111,11 @@ typedef struct VDP_State
 } VDP_State;
 
 void VDP_Init(VDP_State *state);
-void VDP_RenderScanline(VDP_State *state, unsigned short scanline, void (*scanline_rendered_callback)(unsigned short scanline, void *pixels, unsigned short screen_width, unsigned short screen_height));
+void VDP_RenderScanline(VDP_State *state, unsigned int scanline, void (*scanline_rendered_callback)(unsigned int scanline, void *pixels, unsigned int screen_width, unsigned int screen_height));
 
-unsigned short VDP_ReadData(VDP_State *state);
-unsigned short VDP_ReadControl(VDP_State *state);
-void VDP_WriteData(VDP_State *state, unsigned short value);
-void VDP_WriteControl(VDP_State *state, unsigned short value, unsigned short (*read_callback)(void *user_data, unsigned long address), void *user_data);
+unsigned int VDP_ReadData(VDP_State *state);
+unsigned int VDP_ReadControl(VDP_State *state);
+void VDP_WriteData(VDP_State *state, unsigned int value);
+void VDP_WriteControl(VDP_State *state, unsigned int value, unsigned int (*read_callback)(void *user_data, unsigned long address), void *user_data);
 
 #endif /* VDP_H */
